Make macOS bindist configure more deterministic
GHC from a bindist requires running `./configure` and then `make
install`. One would assume that the result is always the same, in
particular if we run this twice on the same host with the same
environment variables. However, this is not the case. See #1461.

The fix is to set an environment variable telling `ar` and friends to
not futz around with timestamps in static library archives. This
environment variable is only necessary and only has an effect on
macOS, but there's no harm in setting it always. If the bindists stop
having a configure stage, or start using `libtool` instead of `ar` and
`ranlib`, then the fix will better be pushed within the bindists
themselves, using the `-D` flag to `libtool`, introduced in XCode
10.2.

Fixes #1461.