Drop support for GHC 8.10.1 and 8.10.2
These two compilers have issues with the way they deal with
passing -no-pie to the linker when the C compiler and/or linker is
overridden via -pgmc and -pgml. Previous compilers do too, but the
workarounds for previous compilers and these compiler versions are
mutually exclusive. In an effort to keep the code simple, we keep the
workaround for previous compiler versions only, and ask all 8.10.*
series users to upgrade to at least 8.10.3.