Add support for GHC 8.10.3
Properly supporting GHC 8.10.3 on all variants of Linux requires
passing -pgmc-supports-no-pie. Otherwise, GHC stops passing -no-pie to
the linker and/or C compiler when it is required, because we override
the C compiler and linker using -pgmc -pgml flags. On earlier GHC,
this flag does not exist. So GHC 8.10.1 and GHC 8.10.2 are simply not
supported by rules_haskell.