Fix handling of DOS-style line endings in cc_wrapper_windows.sh
Sometimes the argument is a response file. In that case, the wrapper
will read the response file line-by-line. But if the response file
uses DOS-style line endings, as GHC >= 8.10 seems intent to do, then
the carriage return character gets embedded in the line, wreaking
havoc to the processing of the argument.