Don't track lib/settings in nixpkgs toolchain
The issue does not occur in the nixpkgs toolchain as the file is always
present in the Nix store, regardless of whether Bazel tracks it.

Tracking it requires exporting it from the `nixpkgs_package` `BUILD`
file for `@rules_haskell_ghc_nixpkgs`. The `BUILD` file defaults to the
template defined in `rules_nixpkgs` and can be overridden by the user,
meaning adding required targets is a breaking change. Given that it is
not required to track the settings file in the nixpkgs use-case, it is
better to avoid the breaking change.