Do not use File objects in dictionary keys
The function `cc_library_key` generates the dictionary keys by which
`LibLibraryToLink` are mapped to their corresponding
`HaskellCcLibraryInfo`. Before this used a `struct` over `None` or
`File` objects. This changes uses `None` or `string` instead, where
`string` is the `.path` attribute of the `File` object.

Using a `File` object in a dictionary key fails if the `File` object
used at lookup time is semantically the same, but has a different
identity due to implementation details in Bazel.

Closes https://github.com/tweag/rules_haskell/issues/1536