Coverage: heuristic for main module name
Coverage is sensible to module naming. Especially, the `Main` module
(i.e. the one with the `main` function) will generate `Main.mix` files.

This forces users to named their file `Main.hs` which is not common
neither convenient in a codebase.

This commit introduces the following heuristics to detect the file which
is the `Main` module:

- Either there is only one file, hence it is the "Main" module.
- it exists a file which is not capitalized, this is the "Main" module,
  because it cannot be an haskell module.

  Close #1593.