Make packages visible to vendored dependents
If a vendored package depends on an otherwise not public package in the
snapshot, then it is made visible to this vendored package.

To handle executable components (tool dependencies) `stack_snapshot`
generates a JSON file listing all components and the `-exe` repository
reads this file to generate the necessary aliases with the correct
visibilities.