Fix c2hs module name detection
Close #1592.

Generated hs files from c2hs are moved to a directory based on the name
of the input file. File `A/B.chs` will generate file `A/B.chs/A/B.hs`.

Module name detection was failing on this context. This fix detects the
presence of `.chs/` content and suppose that it is a `c2hs` generated
file.

This is a bit ad hoc, however it may only fail if we try to compile a
file which appears in a `.chs` directory, which is not to be expected.
(In short, it may fail, but less often that it is currently failing).