Separate _install directory of cabal components
If a cabal package contains multiple libraries, we access them via multiple haskell_cabal_library target.
This results in multiple package databases sharing the same pkgroot.
This is a problem because cabal uses the pkgroot as a package database identifier:
https://github.com/haskell/cabal/blob/master/Cabal/src/Distribution/Simple/Configure.hs#L1995-L2003

This commit changes the install directory of the various components of
cabal packages so that different components are installed in different
pkgroots.