Add versions to `haskell/gen_ghc_bindist.py`
This brings the versions in the generator script and the current versions in `ghc_bindist.bzl`
up to par.

Note, the `upstream` value of `ARCHES` is a list of names now, since for
different GHC versions, there are different supported Debian binary
distributions (e.g. GHC 8.8.4 has deb8 and deb9, but 8.10.1 does not support deb8
anymore).

The first available "upstream" that is available in the hashes is selected.