Don't pass relocatable flag to cabal_wrapper for libs
Two reasons not to do this. First is that Cabal up to 3.6.3.0 doesn't
support relocatable libraries. Second is that there is currently a bug
in Cabal where if you pass this argument and the library uses the
Paths_<pkg> module, Cabal will generate a module that doesn't compile.
This currently prevents building several packages with ghc 9.0 or
ghc 9.2