Avoid :reload in the GHCi script
Since GHC version 9.4.1 a `:reload` in the GHCi script triggers a
secondary recompilation of the modules, effectively doubling GHCi
start-up times.

The `:reload` is required to clear the bindings of the start-up code
that runs to load the modules into scope using the `:module` command.

This logic was needed when rules_haskell didn't know the module names
for modules included in Haskell targets. However, rules_haskell has
changed such that it does now know the module names, at least for
haskell_library, haskell_binary, and haskell_test targets, which are the
only ones relevant for loading sources into the REPL.

This commit changes the GHCi startup script to directly load the modules
based on the module names tracked by rules_haskell. The custom logic in
the startup script is no longer needed, and the `:reload` is no longer
needed since no more bindings need to be cleaned up.