Support using a GHC bindist build with Hadrian
Determine if the current GHC bindist was build by Hadrian. We assume this is the
case if the `config.mk.in` file exists in the root directory of the distribution;
it was located in the `mk` subfolder in older bindists.

We enable a relocatable build, ie. install, by setting `RelocatableBuild` to `YES`
when invoking make. This simplifies the folder structure of the installation.

Also, we no longer need to patch the bin scripts to make them relocatable.