Make stack_snapshot usable in another module's extension
With bzlmod, when using `stack_snapshot` in a module extension of another module,
rules_haskell does not have visibility of generated repositories (such as `rules_haskell_stack`) and cannot build labels to its targets using the `Label` constructor.

To work around this, an optional `label_builder` parameter can be passed to `stack_snapshot`, which is then used to canonicalize various labels.
For the same reason, the `stack_update` attribute to `_stack_snapshot` is now always present (so we do not have to hardcode its label), and is made into a string so that we only fetch the `stack_update` repository when explicitly reading from it.