Make the flags passed to Stack match Cabal
Cabal uses '+' to enable a flag, and '-' to disable a flag. Stack builds
a dictionary that has either 'flag: true', or 'flag: false'. The
behavior before was passing inconsistent flags to the different tools.
Have cabal.bzl handle all the flags that we could pass to cabal. This
helps to aid people new to rules_haskell. It also allows negating a
flag, which didn't work before.