Use a disk cache on windows
Building on windows currently takes a long time (> 2h) and fails eventually when the connection to
the build event upload service is closed.

Using a disk_cache which is stored in Github's action cache hopefully avoids
building stackage libraries every time.

Currently the cache key depends on the hash of the `stackage_snapshot.yaml` file, so it is
evicted whenever that file changes. There is no cache cleanup done currently.