Add protoc wrapper on Windows
This wrapper adds the `bin` folder of the cc toolchain from GHC's bindist to the
`PATH` so that the protoc.exe is able to find its runtime DLLs, e.g. libc++.dll.

Note, using a `sh_binary` rule with a cmd script does not create an extra
executable contrary to using a Bash script.