Remove `libc++.dll.a` to enforce static linking to `libc++`
Newer GHC bindists on Windows come with clang and libc++ instead of gcc and libstdc++.

When linking, `.dll.a` files take precedence over the normal `.a` files and cause the
resulting binary to depend on the corresponding DLL at runtime.

Removing the files causes the linker to pick up the proper static library instead.