Workaround registry URI expansion on Windows
The `%workspace%` placeholder is simply expanded to a windows path like `c:/my/project`.

The resulting URI looks like this: `file:c:/my/project`. This is not a valid file URI and
the `URI.getPath()` method returns `null` which leads to a NPE in Bazel.

Using a triple slash works around this problem, although the path starts with a doubled slash
on unix systems now.