Hide `--registry` flags behind a config
Renovate scans for registry URLs in .bazelrc files and tries to use all of them, but it
does not support file: URLs and fails. However, it ignores any `--registry` values associated
with a configuration.

So introduce a `common` configuration which is always enabled, but causes renovate to
skip the `--registry` values.