Work around `otool` not set to nix store path in GHC settings
In previous GHC versions from nixpkgs, the `otool` setting was referencing a tool in the nix store,
but for GHC 9.6.2 it is just set to "otool" which means it must be in `$PATH`.

The same applies to the `install_name_tool`.

See https://github.com/NixOS/nixpkgs/issues/267250 and https://gitlab.haskell.org/ghc/ghc/-/issues/24211

We work around by using the location of the `ar` command and assume the other tools (from the bintools
package) are also available at the same place.