CI: Introduce "Format & Lint" job
There is no need to build and run buildifier for each module and for every
GHC version in order to keep the code base in shape.

Tag the buildifier tests with `manual` and only run the buildifier for the
top-level repository (which formats / lints all files of the git repo).