Make `find_python` work with Windows App aliases
On recent Windows systems, installing Python makes the python executable available in the `PATH`
but uses a special kind of reparse point which resolves the real executable via a registry lookup.

These reparse points are not supported by the JVM and checking if the file exists will return false.

So, although the file can be executed by running "python" it cannot be found using `repository_ctx.which`.

This change just executes `python3` or `python` and uses the executable information from Python directly
instead of doing a `PATH` lookup.