rules_haskell_tests: Fix ghcide stack pointing to wrong snapshot file
The stack snapshot for GHC 9.2.8 was a symlink to the default snapshot file. So
when no `GHC_VERSION` was set in the environment, it meant the default snapshot
file was used, which is the one for GHC 9.2 but the default GHC version is 9.4.6.

This led to an error on buildkite as `base` in GHC 9.4 was bumped to 4.17:
```
Error: Setup.hs: Encountered missing or private dependencies:
base >=4.8 && <4.17
```

Symlink the default stack snapshot files to the specific version (not the other
way around).