Change start script to point to nixpkgs-23.11-darwin
nixos-23.11 is at end of life, and until we switch to 24.05 or something
newer, we need to use the branch nixpkgs-23.11-darwin instead of
nixos-23.11. CI uses not only the nixpkgs specified in
nixpkgs/default.nix, it also uses the nixpkgs specified by the start
script. When attempting to use this nixpkgs on mac os, the CI runners
were failing due to a build error with python3.11-pytest-xdist-3.3.1.
Since 23.11 is EOL, there won't be any new commits to this branch fixing
the build issue with MacOS.