Fix Update GHC workflow
The workflow failed for a few weeks because of an error:
```
 Traceback (most recent call last):
  File "/home/runner/work/rules_haskell/rules_haskell/haskell/gen_ghc_bindist.py", line 11, in <module>
    from distutils.version import StrictVersion
ModuleNotFoundError: No module named 'distutils'
```

The `distutils` module is deprecated since Python 3.10 and has been removed in Python 3.12 now.

Apparently, this workflow is using Python 3.12 since the `nixpkgs` registry entry by default refers to
"github:NixOS/nixpkgs/nixpkgs-unstable".

This change uses the `packaging.version` module instead, switching to `nix-shell` to make use of the
nixpkgs revision that is configured in `nixpkgs/default.nix` and to provide the `packaging` pip package.