cabal: Add support for empty libraries
Previously, an empty library that only exported modules from another
library failed to build. Bazel would error out when the .a file wasn't
created. These libraries are becoming more common in the Haskell
ecosystem, so we need to support them.

There is a straightforward way to handle these libraries. We make the
package database the default output. We avoid creating an output file
for any libraries if the library is marked as empty. This is all
straightforward bazel dependency/output management.

Adjust the cabal wrapper to fix up the package database even in the case
that there is no library. Otherwise we end up with an empty package
database, which isn't what we're looking for.