cabal: add empty_lib to stack snapshot components
Empty haskell libraries have been a longstanding problem. Add support
for an additional component type `empty_lib`, which is like `lib`,
except that an additional `haskell_cabal_args` target will be created
for the library, instructing Bazel not to look for any object files.

Remove the longstanding package blacklist for empty packages. Replace it
with the now possible correct entries for those libraries, indicating
that the main library is empty.