CI: Update macos image to macos-14
The macos-13 images are being phased out and will stop working soon, see [1].

This means we're switching from x86_64 architecture to arm64 for CI
runs, except that we update to the macos-15-intel runner image for the
examples. This image will be the last available x86_64 image until
August 2027, at which point that architecture will no longer be supported.

[1]: https://github.com/actions/runner-images/issues/13046