Add `non_versioned_output_test` and use it for `test-haddock`
Keeping the expected output files in sync with the different GHC versions and
platforms has always been a tedious task. Also, we did not expect these files
because we knew these would manifest, but only updated the values after a test
failed.

So, this adds a unittest which strips off the version suffixes from the output
files before checking. Now updating / adding a new GHC version should be easier.