Fix ghc-check for GHC 9.10.3
The ghc-check code needs access to the ghc binary in order to check the version.

In previous GHC versions there was a `bin` folder in the libdir; and the bin folder
of the distribution just contained shell scripts calling these binaries.

This has changed now, and the binaries are directly placed into the `bin` folder
and are no longer avaialable inside of `lib`.