cabal: Add static_binary to haskell_cabal_args
Building cabal binaries statically is a fine default, but there are
cases where it would be better to build a dynamically linked binary.
Haskell Language Server is one such example. Certain features are only
available if it is dynamically linked.

Give users a pathway to build dynamically linked binaries if they
choose.